refactor: sms domain을 notification domain으로 리팩토링
