+++ src/main/java/com/rudkids/rudkids/domain/notification/NotificationService.java
@@ -1,9 +1,9 @@
 package com.rudkids.rudkids.domain.notification;
 
-import org.springframework.stereotype.Service;
+import com.rudkids.rudkids.interfaces.notification.dto.NotificationRequest;
 
 public interface NotificationService {
     void sendEmail(String email, String title, String description);
     void sendKakao(String phoneNo, String description);
-    void sendSms(String phoneNo, String description);
+    void sendSms(NotificationRequest.Send request);
 }
