+++ src/main/java/com/rudkids/rudkids/infrastructure/notification/NotificationExecutor.java
@@ -1,14 +1,21 @@
 package com.rudkids.rudkids.infrastructure.notification;
 
 import com.rudkids.rudkids.domain.notification.NotificationService;
+import com.rudkids.rudkids.interfaces.notification.dto.NotificationDtoMapper;
+import com.rudkids.rudkids.interfaces.notification.dto.NotificationRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import net.nurigo.sdk.message.request.SingleMessageSendingRequest;
+import net.nurigo.sdk.message.service.DefaultMessageService;
 import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
 public class NotificationExecutor implements NotificationService {
+    private final DefaultMessageService messageService;
+    private final NotificationDtoMapper notificationDtoMapper;
+
     @Override
     public void sendEmail(String email, String title, String description) {
         log.info("sendEmail");
@@ -20,7 +27,8 @@ public class NotificationExecutor implements NotificationService {
     }
 
     @Override
-    public void sendSms(String phoneNo, String description) {
-        log.info("sendSms");
+    public void sendSms(NotificationRequest.Send request){
+        var message = notificationDtoMapper.toMessage(request);
+        messageService.sendOne(new SingleMessageSendingRequest(message));
     }
 }
