+++ src/main/java/com/rudkids/rudkids/interfaces/notification/dto/NotificationRequest.java
@@ -1,0 +1,12 @@
+package com.rudkids.rudkids.interfaces.notification.dto;
+
+public class NotificationRequest {
+
+    public record Send(
+        String recipientPhoneNumber,
+        String senderPhoneNumber,
+        String title,
+        String content
+    ) {
+    }
+}
