--- src/main/java/com/rudkids/rudkids/interfaces/sms/dto/SmsRequest.java
@@ -1,21 +1,0 @@
-package com.rudkids.rudkids.interfaces.sms.dto;
-
-import net.nurigo.sdk.message.model.Message;
-
-public class SmsRequest {
-
-    public record Send(
-            String recipientPhoneNumber,
-            String senderPhoneNumber,
-            String title,
-            String content) {
-
-        public Message toMessage() {
-            Message message = new Message();
-            message.setFrom(senderPhoneNumber);
-            message.setTo(recipientPhoneNumber);
-            message.setText(title + "\n" + content);
-            return message;
-        }
-    }
-}
